Moved toPrintable() to Utility/String/Ascii.h
