+++ src/Utility/String/Ascii.cpp
@@ -53,4 +53,12 @@ bool noCaseLess(std::string_view a, std::string_view b) {
     return a.size() < b.size();
 }
 
+std::string toPrintable(std::string_view s, char placeholder) {
+    std::string result(s.size(), placeholder);
+    for (size_t i = 0; i < s.size(); i++)
+        if (s[i] >= '\x20' && s[i] <= '\x7E') // Ascii printable chars are in [0x20, 0x7E].
+            result[i] = s[i];
+    return result;
+}
+
 } // namespace ascii
