+++ src/Utility/String/Ascii.h
@@ -27,4 +27,12 @@ struct NoCaseLess {
     }
 };
 
+/**
+ * @param s                             String to transform.
+ * @param placeholder                   Character to replace all non-printable chars with.
+ * @return                              Passed string, with all non-printable non-ascii characters replaced with
+ *                                      `placeholder`.
+ */
+std::string toPrintable(std::string_view s, char placeholder = ' ');
+
 } // namespace ascii
